+++ main/src/main/java/ru/practicum/ewm/events/dto/EventFullDto.java
@@ -24,7 +24,7 @@ public class EventFullDto {
 
     CategoryDto category;
 
-    Integer confirmedRequests;
+    Long confirmedRequests;
 
     @JsonFormat(pattern = DATE_TIME_PATTERN)
     LocalDateTime createdOn;
